Detach Tags into TagsMixin
