+++ gvm/protocols/gmpv214/__init__.py
@@ -96,7 +96,7 @@ from gvm.protocols.gmpv208.entities.severity import (
     SeverityLevel,
     get_severity_level_from_string,
 )
-
+from gvm.protocols.gmpv208.entities.tags import TagsMixin
 from gvm.protocols.gmpv208.entities.tasks import TasksMixin
 from gvm.protocols.gmpv208.entities.tls_certificates import TLSCertificateMixin
 from gvm.protocols.gmpv208.entities.users import (
@@ -146,6 +146,7 @@ class Gmp(
     PortListMixin,
     ReportsMixin,
     ResultsMixin,
+    TagsMixin,
     TargetsMixin,
     TasksMixin,
     TLSCertificateMixin,
